Move `google_results_as_docs` to more appropriate location
